backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
moviepy
librosa
soundfile
numpy
scipy

python-multipart
opencv-python
mediapipe==0.10.14
sqlalchemy
whisper
